requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
atlastk==0.13.3
attrs==24.2.0
certifi==2024.8.30
frozenlist==1.4.1
idna==3.10
iso8601==1.1.0
magic-filter==1.0.12
multidict==6.1.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.2
tortoise-orm==0.21.6
typing_extensions==4.12.2
yarl==1.13.1
